refactor stories structure
