+++ src/stories/autoSize/ByContent.stories.ts
@@ -83,7 +83,7 @@ class LayoutStory
 export const ByContent = (params: any) => new LayoutStory(params);
 
 export default {
-    title: 'Resize',
+    title: 'AutoSize',
     argTypes: argTypes(args),
     args: getDefaultArgs(args)
 };
