+++ src/stories/autoSize/ByParentSize.stories.ts
@@ -55,7 +55,7 @@ class LayoutStory
 export const ByParentSize = (params: any) => new LayoutStory(params);
 
 export default {
-    title: 'Resize',
+    title: 'AutoSize',
     argTypes: argTypes(args),
     args: getDefaultArgs(args)
 };
